correction of ft_memset

+++ 42_files/libft/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: gdornic <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/02 12:12:40 by gdornic           #+#    #+#             */
-/*   Updated: 2023/02/03 13:47:13 by gdornic          ###   ########.fr       */
+/*   Updated: 2023/02/03 13:50:08 by gdornic          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_memset(void *s, int c, size_t n)
 {
 	unsigned char	*pon_s;
-	int		i;
+	size_t		i;
 
 	pon_s = (unsigned char *)s;
 	i = 0;
